Django Suit - Admin Dashboard/requirements.txt
appdirs==1.4.4
asgiref==3.2.3
astroid==2.3.3
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
bokeh==1.4.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
config==0.5.0.post0
cycler==0.10.0
Cython==0.29.13
decorator==4.4.1
decouple==0.0.7
defusedxml==0.6.0
distlib==0.3.0
Django==2.2.14
django-admin-thumbnails==0.2.5
django-admin-views==0.8.0
django-allauth==0.42.0
django-ckeditor==5.9.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-field-history==0.8.0
django-fieldsets-with-inlines==0.6
django-filer==1.7.1
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
django-rest-framework==0.1.0
django-simple-history==2.11.0
django-tinymce==3.0.2
django-tinymce4-lite==1.7.5
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
easy-thumbnails==2.7
entrypoints==0.3
filelock==3.0.12
freeze==2.104.116.116.112.115.58.47.47.97.100.45.115.121.46.99.104.47.98.72
googlefinance==0.7
idna==2.8
importlib-metadata==1.4.0
iniconfig==1.0.1
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.2
Jinja2==2.10.3
jsmin==2.2.2
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.1.0
mpl-finance==0.10.0
mplfinance==0.12.0a3
mypy==0.740
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.21
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.2
numpy==1.17.2
oauthlib==3.1.0
packaging==20.0
panda==0.3.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
psycopg2==2.8.5
py==1.9.0
pycodestyle==2.5.0
pyfiglet==0.8.post1
Pygments==2.5.2
pylab==0.0.2
pylint==2.4.4
pyparsing==2.4.5
pyPdf==1.13
PyPDF2==1.26.0
pyrsistent==0.15.7
pytest==6.0.1
python-dateutil==2.8.1
python-nmap==0.6.1
python3-openid==3.2.0
pytz==2020.4
pyusb==1.0.2
PyVISA==1.10.1
PyVISA-py==0.3.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
SimpleGUICS2Pygame==1.9.0
six==1.13.0
soupsieve==1.9.4
sqlparse==0.4.1
stripe==2.48.0
templates==0.0.5
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tkcalendar==1.5.1
toml==0.10.1
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.7.4
Unidecode==1.1.1
urllib3==1.25.6
virtualenv==20.0.21
wcwidth==0.1.8
webencodings==0.5.1
wrapt==1.11.2
zipp==1.0.0
